fix some bugs when send small message

+++ comm_network/epoll/epoll_comm_network.h
@@ -6,7 +6,7 @@ class EpollCommNet final : public CommNet {
  public:
   EpollCommNet() {}
   ~EpollCommNet() {}
-  void SendMsg(int64_t dst_machine_id, int64_t msg_type, const char* ptr, size_t bytes,
+  void SendMsg(int64_t dst_machine_id, int32_t msg_type, const char* ptr, size_t bytes,
                std::function<void()> handler = NULL) {}
   void DoRead(int64_t src_machine_id, void* src_addr, size_t bytes, void* dst_addr,
               std::function<void()> handler = NULL) {}
